Sacrifice small compresion efficiency at level 1,2 and 3 for a speedup on material that is hard to compress.
If we made no matches on level 1+2+3, do not attempt to huffman compress the data, but store it directly instead.

If we reduced the size by less than 10 bytes through matching, huffman compress the data.

e.txt is a worst case scenario, since it cannot be compressed through matching, but a lot through huffman compression.