zstd: Add WithEncoderPadding
WithEncoderPadding will add padding to all output so the size will be a multiple of n.
This can be used to obfuscate the exact output size.
The contents will be a skippable frame, so it will be invisible by the decoder.
The padded area will be filled with data from crypto/rand.Reader.
'n' must be > 0 and <= 1GB, 1<<30 bytes.
If `EncodeAll` is used with data already in the destination, the total size will be multiple of this.