inflate: Read more bits when decoding
Fixes #231

When a block is not marked as the final block it should be possible to read 10 further bits ahead.

The smallest block seems to be a predefined block with a single EOB, which would be 10 bits + EOB from current block (current limit).

This should make it possible to fill more bits at the time when decoding.