[skip ci] Trying to compile decodeSync with avo
After spilling to the stack all execute-related values,
we're still running out of registers. Need to figure out
which values from decode can be spilled too.

However, I'm unhappy with this, as my initial version of
decodeSync didn't use any stack. I believe it can be done,
but I'd like to share as many code as possible. On the other
hand, having code full of short if-statements is not readable.