readerWrapper: readByte needs to use io.ReadFull
This fixes an "unexpected EOF" error that can occur when streaming
from a network source into the zstd decompressor.

Even when asking for one byte of data, it's necessary to use
io.ReadFull. It's valid to return 0 bytes without EOF on a read,
and this behavior is seen on network sources rather frequently.