s2: add GetBufferCapacity() method (#832)
Add GetBufferCapacity() method. We are reusing readers with sync.Pool
and we'd like to avoid allocating memory for the default block size
since most of the inputs are smaller. To have a better estimate of how
big the lazy buffer should be, we are thinking about keeping in mind a
running average of the internal buffer capacities. This method would
allow us to implement that.