Fix invalid encoding on level 9 with single value input (#1115)
* Fix invalid encoding on level 9 with single value input

With single value input and a full block write (>=64K) the indexing function would overflow a uint16 to a 0.

This would make it impossible to generate a valid huffman table for the literal size prediction.

In turn this would mean that the entire block would be output as literals - since the cost of the value would be 0 bits.

This would in turn mean that EOB could not be encoded for the bit writer - since there were no matches. This was previously being satisfied with "filling".

Fixes:

1. First never encode more than `maxFlateBlockTokens` - 32K for the literal estimate table.
2. Always include EOB explicitly - if somehow literals should slip through.
3. Add test that will write big single-value input as regression test. Others were using copy that does smaller writes.

Fixes #1114

* Retract v1.18.1